+++ django_afip/pdf.py
@@ -7,12 +7,16 @@ from io import BytesIO
 from typing import TYPE_CHECKING
 
 import qrcode
+from django_renderpdf.helpers import render_pdf
 
 if TYPE_CHECKING:
+    from typing import IO
+
     from PIL import Image
 
     from django_afip.models import Receipt
     from django_afip.models import ReceiptPDF
+
 
 logger = logging.getLogger(__name__)
 
@@ -70,3 +74,104 @@ def get_encoded_qrcode(receipt_pdf: ReceiptPDF) -> str:
     qr_img.save(img_data, format="PNG")
 
     return base64.b64encode(img_data.getvalue()).decode()
+
+
+# Note: When updating this, be sure to update the docstring of the method that uses
+# these below.
+TEMPLATE_NAMES = [
+    "receipts/{taxpayer}/pos_{point_of_sales}/code_{code}.html",
+    "receipts/{taxpayer}/code_{code}.html",
+    "receipts/code_{code}.html",
+    "receipts/{code}.html",
+]
+
+
+class PdfBuilder:
+    """Builds PDF files for Receipts.
+
+    Creating a new instance of a builder does nothing; use :meth:`~render_pdf` to
+    actually render the file.
+
+    This type can be subclassed to add custom behaviour or data into PDF files.
+    """
+
+    def __init__(self) -> None:
+        pass
+
+    def get_template_names(self, receipt: Receipt) -> list[str]:
+        """Return the templates use to render the Receipt PDF.
+
+        Template discovery tries to find any of the below receipts::
+
+            receipts/{taxpayer}/pos_{point_of_sales}/code_{code}.html
+            receipts/{taxpayer}/code_{code}.html
+            receipts/code_{code}.html
+            receipts/{code}.html
+
+        To override, for example, the "Factura C" template for point of sales 0002 for
+        Taxpayer 20-32964233-0, use::
+
+            receipts/20329642330/pos_2/code_6.html
+        """
+        return [
+            template.format(
+                taxpayer=receipt.point_of_sales.owner.cuit,
+                point_of_sales=receipt.point_of_sales.number,
+                code=receipt.receipt_type.code,
+            )
+            for template in TEMPLATE_NAMES
+        ]
+
+    def get_context(self, receipt: Receipt) -> dict:
+        """Returns the context used to render the PDF file."""
+        from django_afip.models import Receipt
+        from django_afip.models import ReceiptPDF
+
+        context: dict = {}
+
+        receipt_pdf = (
+            ReceiptPDF.objects.select_related(
+                "receipt",
+                "receipt__receipt_type",
+                "receipt__document_type",
+                "receipt__validation",
+                "receipt__point_of_sales",
+                "receipt__point_of_sales__owner",
+            )
+            .prefetch_related(
+                "receipt__entries",
+            )
+            .get(receipt=receipt)
+        )
+
+        # Prefetch required data in a single query:
+        receipt_pdf.receipt = (
+            Receipt.objects.select_related(
+                "receipt_type",
+                "document_type",
+                "validation",
+                "point_of_sales",
+                "point_of_sales__owner",
+            )
+            .prefetch_related(
+                "entries",
+            )
+            .get(
+                pk=receipt_pdf.receipt_id,
+            )
+        )
+        taxpayer = receipt_pdf.receipt.point_of_sales.owner
+
+        context["pdf"] = receipt_pdf
+        context["taxpayer"] = taxpayer
+        context["qrcode"] = get_encoded_qrcode(receipt_pdf)
+
+        return context
+
+    def render_pdf(self, receipt: Receipt, file_: IO) -> None:
+        """Renders the PDF into ``file_``."""
+        render_pdf(
+            template=self.get_template_names(receipt),
+            file_=file_,
+            context=self.get_context(receipt),
+        )